Fix compilation errors after merge in PlayerEventHandler

+++ src/main/java/cat/udl/eps/entsoftarch/webingogeiadeapi/handler/PlayerEventHandler.java
@@ -57,19 +57,16 @@ public class PlayerEventHandler {
             System.out.println(player.toString());
         }
         else if (player.getFromWallet() != 0) {
-            int wallet = player.getWallet();
+            double wallet = player.getWallet();
             int value = player.getFromWallet();
             if (value < 5) {
                 throw new DepositMoneyException("Not enought money");
             }
             player.setWallet(wallet - value);
             player.setFromWallet(0);
-            System.out.println(player.toString());        }
-
-        // playerRepository.save(player);
+            System.out.println(player.toString());
+        }
     }
-
-
 
     @HandleBeforeDelete
     public void handlePlayerPreDelete(Player player){
